Add OpenFOAM's control dictionary to the image
The control dict is used by the simpleFoam program to load default settings for the FOAM processor. Environment variable WM_PROJECT_DIR must point to the parent folder of the etc dir where controlDict can be found. This seems to be hardcoded in OpenFOAM so it is not possible to change.