Support OpenFOAM input case upload to a pre-built image
We need to pass new case pathname, new image pathname and
optionally OSv home folder if it is not specified as a env
variable $OSV_HOME. The base image is first copied then a
new case is uploaded to the copy. Finnally we run the copy
with  $OSV_HOME/script/run -i <new image path>. Nice thing
is that our new image can be anywhere.
The case upload at build time is disabled to prevent overriding
and other potential problems. The case can be uploaded after
build time.

Change-Id: I89c61ecc48f2fb12807efcf47d4df00864df0a18
Signed-off-by: ales <ales.omerzel@xlab.si>