Provide a simple example of OpenMP
OpenMP is a library that is used in the Bones use case. This application is a
simple example of an OpenMP enabled application that should result in a number
of threads greeting the world. The number of threads should be the same as the
number of available cores, but can be modified via OMP_NUM_THREADS environment
variable.

Change-Id: Iea9a5113ff2017cf7f861b87b355d461b462783d
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>