Add a mockup app for running Java subprocesses
This is a demo application replicating the important parts of the
workflow Apache Storm uses when executing Storm workers (spouts and
bolts) from the Supervisor process. The purpose of this demo app is to
have a simple application suitable for testing of the new OSv Java API
to launch "subprocesses" as threads with separate environments.

Refer to README for instructions on how to use.

Change-Id: If843c2bbfd61cf9d0d48fadb07b8b4679b86d963
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>