OpenFOAM: fix libfluidThermophysicalModels missing dependency
libfluidThermophysicalModels.so uses functions from libspecie.so.
The dependency library info is not included into the compiled
binary, as visible in ldd output. For example, _ZN4Foam6specie4TstdE
(Foam::specie::Tstd) is one of the missing functions.

This causes problem in OSv, when rhoSimpleFoam/rhoPorousSimpleFoam
crash because of loading libfluidThermophysicalModels before
libspecie.

Change-Id: Iab53e8a773e950620aba8dd2d98ef0ef07dbcd19
Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>