Open MPI: change usr.manifest generation code
The libibverbs.so.1 can be in /usr/lib (ubuntu) or /usr/lib64 (centos).
Also, other libs are missing (libnuma).
This will be fixed in follow-up commit; that commit is only preparation.

There is only one change in the generated usr.manifest. Help files in
OSv image are put in $HOME/openmpi-bin/share/openmpi/ directory. This is
required, as we run Open MPI configure with --prefix=$HOME/openmpi-bin,
and that path becomes hardcoded/baked in compiled binaries. Previously,
we put them into /usr/local/share/openmpi, and they were not found - but
that was visible only when some runtime error occured (say out-of-memory
etc).

Change-Id: I562e52da634f975c5c8f3dd5fbbf60bff2911b1c
Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>