Open MPI: check for build dependencies
Check if libnuma and libibvers development libraries are installed
on the build machine. If not, show message and abort build.

In order to use CPU and memory affinity, Open MPI has to be compiled
with hwloc and libnuma support. The ./configure stage will disable
hwloc/libnuma if libnuma is not found.

ibverbs will be required for vRDMA usage.

Change-Id: I7af68e42d90a728f5f7a69b00bc7c1c444db0b64
Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>