Do not set the default run command
OpenFOAM app is usually bundled along with the CLI module that sets it's
own default command line (`/cli/cli.so`). Previously, OpenFOAM app would
override this causing the display of the  help message of the
`simpleFoam` solver and immediate termination of the VM.

This patch removes the default command. Some examples are still given in
the `module.py` file, but need to be manually enabled by the user of
this app.

Change-Id: Id073bf13f70c5f13c621629890ee4fb80077cbca
Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>